refactor: Merged user with accounts endpoint
