--- backend/src/routes/account/mod.rs
@@ -1,5 +1,0 @@
-mod delete;
-
-pub fn config(cfg: &mut utoipa_actix_web::service_config::ServiceConfig) {
-    cfg.service(delete::delete);
-}
