+++ backend/src/routes/mod.rs
@@ -1,4 +1,3 @@
-pub mod account;
 pub mod auth;
 pub mod event;
 pub mod index;
